+++ lib/src/widgets/widgets.dart
@@ -1,1 +1,8 @@
-export 'surface.dart';
+export 'elevation/surface.dart';
+export 'motion/container_transform.dart';
+export 'motion/container_transform_route.dart';
+export 'motion/fade.dart';
+export 'motion/fade_through.dart';
+export 'motion/page_transitions.dart';
+export 'motion/shared_axis_transition.dart';
+export 'title.dart';